style restructure
