+++ projects/ui-framework/src/lib/form-elements/form-elements.scss
@@ -1,0 +1,122 @@
+@import '../style/colors';
+
+$form-field-height: 46px;
+$textarea-height: 120px;
+
+@mixin MatFormField() {
+  .mat-form-field {
+    width: 100%;
+    margin: 20px 0;
+    font-family: inherit;
+    line-height: 1;
+    .mat-form-field-wrapper {
+      padding: 0;
+      .mat-form-field-flex {
+        border: 1px solid $grey-500;
+        box-sizing: border-box;
+        background-color: $white;
+        height: auto;
+        border-radius: 4px;
+        position: relative;
+        place-content: center;
+        align-items: center;
+        transition: box-shadow 0.2s;
+        .mat-form-field-infix {
+          padding: 0;
+          border: none;
+          height: 100%;
+          position: relative;
+          input, textarea, mat-select {
+            margin: 0;
+            box-sizing: border-box;
+            height: $form-field-height;
+            padding: 0 16px;
+            &[type=number]::-webkit-inner-spin-button,
+            &[type=number]::-webkit-outer-spin-button {
+              -webkit-appearance: none;
+              margin: 0;
+            }
+          }
+          .mat-form-field-label-wrapper {
+            box-sizing: border-box;
+            top: 0;
+            padding: 0;
+            display: flex;
+            height: 100%;
+            place-content: center space-between;
+            align-items: center;
+            overflow: visible;
+            .mat-form-field-label {
+              box-sizing: border-box;
+              padding: 0 16px;
+              display: block;
+              top: auto;
+              position: relative;
+              color: $grey-600;
+              transition: transform 0.2s;
+            }
+          }
+        }
+        .mat-form-field-prefix, .mat-form-field-suffix {
+          padding: 0 5px;
+        }
+      }
+      .mat-form-field-subscript-wrapper {
+        top: auto;
+        margin-top: 4px;
+        font-size: 12px;
+        padding: 0 16px;
+      }
+      .mat-form-field-underline {
+        display: none;
+      }
+    }
+    &:hover {
+      .mat-form-field-wrapper .mat-form-field-flex {
+        box-shadow: 2px 2px 3px 0 rgba(0, 0, 0, 0.04);
+      }
+    }
+    &.mat-focused {
+      .mat-form-field-wrapper .mat-form-field-flex {
+        border: 1px solid $grey-600;
+        box-shadow: none;
+      }
+    }
+    &.mat-form-field-should-float {
+      .mat-form-field-wrapper .mat-form-field-flex .mat-form-field-infix .mat-form-field-label-wrapper .mat-form-field-label {
+        transform: translateY(-34px);
+      }
+    }
+    &.mat-form-field-disabled {
+      .mat-form-field-wrapper .mat-form-field-flex {
+        background-color: $grey-100;
+        .mat-form-field-infix .mat-form-field-label-wrapper .mat-form-field-label {
+          color: $grey-500;
+        }
+      }
+    }
+    &.required {
+      .mat-form-field-wrapper .mat-form-field-flex .mat-form-field-infix .mat-form-field-label-wrapper .mat-form-field-label {
+        &:after {
+          content: '*';
+          display: inline-block;
+          margin-left: 2px;
+        }
+      }
+    }
+    &.error {
+      .mat-form-field-wrapper .mat-form-field-flex {
+        border: 1px solid var(--negative-500);
+      }
+      .mat-form-field-subscript-wrapper .mat-hint {
+        color: var(--negative-500);
+      }
+    }
+    &.hide-label-on-focus.mat-focused {
+      .mat-form-field-wrapper .mat-form-field-flex .mat-form-field-infix .mat-form-field-label-wrapper .mat-form-field-label {
+        display: none;
+      }
+    }
+  }
+
+}
