+++ projects/ui-framework/src/lib/form-elements/input/input.component.scss
@@ -1,5 +1,6 @@
-@import '../../style/colors';
+@import '../form-elements.scss';
 
 :host ::ng-deep {
   display: block;
+  @include MatFormField();
 }
