+++ projects/ui-framework/src/lib/form-elements/select-elements/select/select.component.scss
@@ -1,13 +1,25 @@
 @import '../../../style/typography';
 @import '../../../style/colors';
-@import '../../../style/mat-form-field';
+@import '../../form-elements.scss';
 
 :host ::ng-deep {
   @include BodyFont(normal, normal, 12px);
-  .mat-select-trigger {
-    height: 100%;
-    .mat-select-value {
-      vertical-align: middle;
+  @include MatFormField();
+  .mat-form-field {
+    &.mat-focused {
+      .mat-form-field-wrapper .mat-form-field-flex {
+        border-bottom-right-radius: 0;
+        border-bottom-left-radius: 0;
+        border: 1px solid $grey-500;
+      }
+    }
+    .mat-form-field-wrapper .mat-form-field-flex {
+      .mat-select-trigger {
+        height: 100%;
+        .mat-select-value {
+          vertical-align: middle;
+        }
+      }
     }
   }
 }
