+++ projects/ui-framework/src/lib/overlay/overlay.scss
@@ -1,0 +1,74 @@
+@import '../style/colors.scss';
+@import '../style/shapes.scss';
+
+@mixin OverlayPanel() {
+  position: relative;
+  overflow: visible;
+  border: 1px solid $grey-500;
+  box-shadow: 0 2px 5px 0 rgba(0, 0, 0, 0.05);
+  border-radius: 4px;
+  box-sizing: border-box;
+}
+
+@mixin OverlayPointerHolders() {
+  z-index: 1;
+
+  &:before, &:after {
+    content: '';
+    display: block;
+    position: absolute;
+  }
+
+  &:before {
+    z-index: 0;
+  }
+  &:after {
+    z-index: 2;
+  }
+}
+
+@mixin OverlayPositionBounding($direction) {
+  @if $direction == "below" {
+    top: 8px;
+    &:before {
+      top: -5px;
+      @include arrow-up(7px, $grey-500);
+      border-bottom-width: 5px;
+    }
+    &:after {
+      top: -4px;
+      @include arrow-up(5px, $white);
+    }
+  }
+
+  @if $direction == "above" {
+    bottom: 8px;
+    &:before {
+      bottom: -5px;
+      @include arrow-down(7px, $grey-500);
+      border-top-width: 5px;
+    }
+    &:after {
+      bottom: -4px;
+      @include arrow-down(5px, $white);
+    }
+  }
+
+  @if $direction == "after" {
+    &:before {
+      left: 11px;
+    }
+    &:after {
+      left: 13px;
+    }
+  }
+
+  @if $direction == "before" {
+    &:before {
+      right: 11px;
+    }
+    &:after {
+      right: 13px;
+    }
+  }
+}
