+++ projects/ui-framework/src/lib/overlay/panel/panel.component.scss
