+++ projects/ui-framework/src/lib/overlay/panel/panel.module.ts
