+++ projects/ui-framework/src/lib/style/style.scss
@@ -1,7 +1,5 @@
 @import './variables.scss';
 @import './typography.scss';
-@import './mat-form-field.scss';
-@import './mat-menu-panel.scss';
 
 body {
   font-family: var(--body-font-family);
